cross-account-region-data-pipeline/python/requirements.txt
aws_ddk_core==1.0.0
constructs>=10.0.5
aws-cdk-lib==2.71.0

